requirements-local.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.32

gunicorn==23.0.0
python-dotenv==1.0.1
openai==1.40.2
httpx==0.27.2
requests==2.32.3
python-slugify==8.0.4

